spreadOut

Prepares input mass spectrometry data for subsequent functions. Relevant information for every instance of every chemical across all samples is placed on the same matrices with no overlaps. Chemicals are also searched on PubChem and all published names, published m/z histograms, and exact mass data are downloaded and stored as a large list.

spreadOut(input)

Arguments

input

A .CSV file containing raw mass spectrometry output with 'Component.RT', 'Component.Area', 'Base.Peak.MZ', 'File.Name', 'Compound.Name', and 'Match.Factor' as column names/required input

Value

Returns a list with matrices storing raw area values, chemical identities, m/z values, match factors, retention times, exact mass data (if published), and a unique code for each data point (retention time pasted to exact mass). The list also contains an additional list with all published chemical names, top m/z peaks, exact mass, and structural fingerprints in SDF format.

Details

Automatically prepares .CSV input containing retention time, detected m/z, component area, compound name, match factor, and sample or file name for downstream uafR functions.

Examples

spreadOut(standard_data)
#> Welcome to uafR! Preparing your data, please be patient.
#> 
#> [Current/Total] |--Exact Mass--|--CMP Name 1--|--CMP Name 2--|--Top MZ Peaks--|
#> [1 / 63]: 68.0045147 | trifluoroborane | Boron trifluoride | 49 | 30 | 43
#> Warning: URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/71583/JSON?heading=Depositor-Supplied+Synonyms': Timeout of 60 seconds was reached
#> [3 / 63]: 83.9533555 | dichloromethane | Methylene chloride | 49 | 57.85 | 84 | 36.56 | 51
#> Warning: cannot open URL 'https://cactus.nci.nih.gov/chemical/structure/6-Phenyl-6H-1,2,3,4,5,6,7-[1,2,3]triazolo[4,5-e][1,2,3,4]tetrazine-1,3-dione/smiles': HTTP status was '500 Internal Server Error'
#> Warning: cannot open URL 'https://cactus.nci.nih.gov/chemical/structure/6-Phenyl-6H-1,2,3,4,5,6,7-[1,2,3]triazolo[4,5-e][1,2,3,4]tetrazine-1,3-dione/stdinchikey': HTTP status was '500 Internal Server Error'
#> [4 / 63]: NA | 2-propanone | 6-Phenyl-6H-1,2,3,4,5,6,7-[1,2,3]triazolo[4,5-e][1,2,3,4]tetrazine-1,3-dione | 43 | 30 | 7.85 | 42 | 23.21
#> Warning: NAs introduced by coercion
#> Warning: NAs introduced by coercion
#> [5 / 63]: 123.032028402 | nitrobenzene | Benzene, nitro- | 77 | 12.44 | 51 | 54.26 | 50
#> [6 / 63]: 70.016712692 | 1,2,5-oxadiazole | 1,2,5-Oxadiazole | 43 | 13 | 18
#> [7 / 63]: 128.03342538 | 3-amino-1H-1,2,4-triazole-5-carboxylic acid | 3-Amino-1,2,4-triazole-5-carboxylic acid | 18 | 31 | 57 | 13 | 11
#> Warning: URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug/compound/cid/9571063/SDF': Timeout of 60 seconds was reached
#> Error in mass_info@SDF: no applicable method for `@` applied to an object of class "character"