categorate

Searches a vector of chemical names on PubChem for published information. Uses regular expressions to extract information from 6 data- bases: LOTUS the natural products occurrence database [LOTUS](https://lotus.naturalproducts.net/), Flavor and Extract Manufacturers Association [FEMA](https://www.femaflavor.org/), Kyoto Encyclopedia of Genes and Genomes [KEGG](https://www.genome.jp/kegg/), Food and Drug Administration [FDA/SPL](https://www.fda.gov/), and Reactive Groups from [PubChem](https://pubchem.ncbi.nlm.nih.gov/). Also downloads structural data in SDF format to summarize the atomic structures and common molecular groups of each chemical. Finally, the function downloads structural data for the chemicals within each group of the input library (`chemical_library`). Each chemical of interest (`compounds`) is tested against every group of chemicals for structural overlaps that exceed a threshold of 0.85.

categorate(compounds, chemical_library, input_format = "wide")

Arguments

compounds

A vector containing chemical names in IUPAC notation (preferred).

chemical_library

A .CSV file containing columns with chemical groups. Column names label the group while the rows contain chemicals that are described by the label.

input_format

Designates the data structure for input library (`chemical_library`). Default setting is "wide" can be changed to "long."

Value

List with 4 data frames: (1) information from database searches, (2) atomic and molecular sub-group structure summary, (3) strong (similarity > 0.95) or moderate (similarity > 0.85) matches with input library groups, and (4) top chemical from groups a chemical shared a strong match with.

Details

Provides a detailed overview of categorical and structural information for every chemical of interest. Functional matches are also generated by matching with groups from a chemical library.

Examples

compounds = c("3-Octanone","Decane","Mesitylene","1,2,4-trimethyl-benzene",
"D-Limonene","beta-ethyl benzeneethanol","1,4-diethyl benzene",
"1,2-diethyl benzene","1,3,8-p-Menthatriene","(2-methyl-1-propenyl)-Benzene",
"1-Phenyl-1-butene","Linalool","Nonanal","5-nonyl-2-Thiophenecarboxylic acid",
"Dichloroacetaldehyde","Linalyl acetate","Beta-Ocimene")
categorate(compounds, library_data, input_format = "wide")
#> Searching query chemicals for structural data, please be patient!
#> [1/17]-3-Octanone
#> [2/17]-Decane
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/15600/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/15600/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> [3/17]-Mesitylene
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/7947/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> [4/17]-1,2,4-trimethyl-benzene
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/7247/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/7247/JSON?source=FDA/SPL+Indexing+Data': HTTP status was '404 Not Found'
#> [5/17]-D-Limonene
#> [6/17]-beta-ethyl benzeneethanol
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/95542/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/95542/JSON?source=LOTUS+-+the+natural+products+occurrence+database': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/95542/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/95542/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/95542/JSON?source=Medical+Subject+Headings+(MeSH)': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/95542/JSON?source=FDA/SPL+Indexing+Data': HTTP status was '404 Not Found'
#> [7/17]-1,4-diethyl benzene
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/7734/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/7734/JSON?source=LOTUS+-+the+natural+products+occurrence+database': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/7734/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> [8/17]-1,2-diethyl benzene
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/8657/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/8657/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/8657/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> [9/17]-1,3,8-p-Menthatriene
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/176983/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/176983/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/176983/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/176983/JSON?source=Medical+Subject+Headings+(MeSH)': HTTP status was '404 Not Found'
#> [10/17]-(2-methyl-1-propenyl)-Benzene
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/13030/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/13030/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/13030/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/13030/JSON?source=Medical+Subject+Headings+(MeSH)': HTTP status was '404 Not Found'
#> [11/17]-1-Phenyl-1-butene
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/123088/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/123088/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/123088/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/123088/JSON?source=FDA/SPL+Indexing+Data': HTTP status was '404 Not Found'
#> [12/17]-Linalool
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/6549/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/6549/JSON?source=LOTUS+-+the+natural+products+occurrence+database': Timeout of 60 seconds was reached
#> [13/17]-Nonanal
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/31289/JSON?source=KEGG': HTTP status was '404 Not Found'
#> [14/17]-5-nonyl-2-Thiophenecarboxylic acid
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/587962/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/587962/JSON?source=LOTUS+-+the+natural+products+occurrence+database': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/587962/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/587962/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/587962/JSON?source=Medical+Subject+Headings+(MeSH)': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/587962/JSON?source=FDA/SPL+Indexing+Data': HTTP status was '404 Not Found'
#> [15/17]-Dichloroacetaldehyde
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/6576/JSON?source=LOTUS+-+the+natural+products+occurrence+database': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/6576/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> [16/17]-Linalyl acetate
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/8294/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/8294/JSON?source=LOTUS+-+the+natural+products+occurrence+database': Timeout of 60 seconds was reached
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/8294/JSON?source=KEGG': HTTP status was '404 Not Found'
#> [17/17]-Beta-Ocimene
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/18756/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/18756/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/18756/JSON?source=FDA/SPL+Indexing+Data': HTTP status was '404 Not Found'
#> Setting up your input library for Tanimoto matching...
#> [1/5]-Type.A
#> [1/4]-2-Aminothiazole
#> [2/4]-3,4,5,6-Tetrachlorocyclohexene
#> [3/4]-N-methyl-1,3,5-triazin-2-amine
#> [4/4]-2-Methyloctahydro-2-azacyclopropa[cd]pentalene
#> [2/5]-Type.B
#> [1/6]-o-Cresol
#> [2/6]-Salicylic Acid
#> [3/6]-Guaiacol
#> [4/6]-Aspirin
#> [5/6]-Salicyl alcohol
#> [6/6]-4-Methylsalicylic acid
#> [3/5]-Type.C
#> [1/3]-Nonane
#> [2/3]-Dodecane
#> [3/3]-Tridecane
#> [4/5]-Type.D
#> [1/7]-Butyl methacrylate
#> [2/7]-Isobutyl hexanoate
#> [3/7]-Ethyl heptanoate
#> [4/7]-Methyl hexanoate
#> [5/7]-Methyl heptanoate
#> [6/7]-Dihexyl adipate
#> [7/7]-2-Heptanone
#> [5/5]-Type.E
#> [1/7]-Octane
#> [2/7]-Octanoic acid
#> [3/7]-1-Octanol
#> [4/7]-Hexadecanal
#> [5/7]-Decanal
#> [6/7]-Undecanal
#> [7/7]-Hexyl acetate
#> [1] "No duplicates detected!"
#> $reactives
#>                              reactives                           Chemical
#> 1                              Ketones                         3-Octanone
#> 2                  ETHYL N-AMYL KETONE                         3-Octanone
#> 3                               Decane                             Decane
#> 4    Hydrocarbons, Aliphatic Saturated                             Decane
#> 5                             N-DECANE                             Decane
#> 6               Hydrocarbons, Aromatic                         Mesitylene
#> 7               Hydrocarbons, Aromatic            1,2,4-trimethyl-benzene
#> 8                       Limonene, (+)-                         D-Limonene
#> 9  Hydrocarbons, Aliphatic Unsaturated                         D-Limonene
#> 10                                None          beta-ethyl benzeneethanol
#> 11                                None                1,4-diethyl benzene
#> 12                                None                1,2-diethyl benzene
#> 13                                None               1,3,8-p-Menthatriene
#> 14                                None      (2-methyl-1-propenyl)-Benzene
#> 15                                None                  1-Phenyl-1-butene
#> 16                                None                           Linalool
#> 17                           Aldehydes                            Nonanal
#> 18                                None 5-nonyl-2-Thiophenecarboxylic acid
#> 19                Dichloroacetaldehyde               Dichloroacetaldehyde
#> 20                           Aldehydes               Dichloroacetaldehyde
#> 21       Halogenated Organic Compounds               Dichloroacetaldehyde
#> 22             Polymerizable Compounds               Dichloroacetaldehyde
#> 23                                None                    Linalyl acetate
#> 24                                None                       Beta-Ocimene
#> 
#> $LOTUS
#>         LOTUS                           Chemical
#> 1   Q18349104                         3-Octanone
#> 2     biochem                         3-Octanone
#> 3     Q150717                             Decane
#> 4     biochem                             Decane
#> 5     Q425161                         Mesitylene
#> 6     biochem                         Mesitylene
#> 7     Q376994            1,2,4-trimethyl-benzene
#> 8     biochem            1,2,4-trimethyl-benzene
#> 9   Q27888324                         D-Limonene
#> 10    biochem                         D-Limonene
#> 11       None          beta-ethyl benzeneethanol
#> 12       None                1,4-diethyl benzene
#> 13   Q2429316                1,2-diethyl benzene
#> 14    biochem                1,2-diethyl benzene
#> 15  Q27161428               1,3,8-p-Menthatriene
#> 16    biochem               1,3,8-p-Menthatriene
#> 17  Q81989319      (2-methyl-1-propenyl)-Benzene
#> 18    biochem      (2-methyl-1-propenyl)-Benzene
#> 19 Q105169590                  1-Phenyl-1-butene
#> 20    biochem                  1-Phenyl-1-butene
#> 21       None                           Linalool
#> 22    Q419668                            Nonanal
#> 23    biochem                            Nonanal
#> 24       None 5-nonyl-2-Thiophenecarboxylic acid
#> 25       None               Dichloroacetaldehyde
#> 26       None                    Linalyl acetate
#> 27  Q27108635                       Beta-Ocimene
#> 28    biochem                       Beta-Ocimene
#> 
#> $KEGG
#>                             KEGG                           Chemical
#> 1                    KEGG: Lipid                         3-Octanone
#> 2                           <NA>                         3-Octanone
#> 3                           <NA>                         3-Octanone
#> 4                           None                             Decane
#> 5                           None                         Mesitylene
#> 6                           None            1,2,4-trimethyl-benzene
#> 7                    KEGG: Lipid                         D-Limonene
#> 8  KEGG: Phytochemical Compounds                         D-Limonene
#> 9                           <NA>                         D-Limonene
#> 10                          <NA>                         D-Limonene
#> 11                          <NA>                         D-Limonene
#> 12                          None          beta-ethyl benzeneethanol
#> 13                          None                1,4-diethyl benzene
#> 14                          None                1,2-diethyl benzene
#> 15                          None               1,3,8-p-Menthatriene
#> 16                          None      (2-methyl-1-propenyl)-Benzene
#> 17                          None                  1-Phenyl-1-butene
#> 18                   KEGG: Lipid                           Linalool
#> 19 KEGG: Phytochemical Compounds                           Linalool
#> 20                          <NA>                           Linalool
#> 21                          <NA>                           Linalool
#> 22                          <NA>                           Linalool
#> 23                          None                            Nonanal
#> 24                          None 5-nonyl-2-Thiophenecarboxylic acid
#> 25                          None               Dichloroacetaldehyde
#> 26                          None                    Linalyl acetate
#> 27                          None                       Beta-Ocimene
#> 
#> $FEMA
#>         FEMA                           Chemical
#> 1     Butter                         3-Octanone
#> 2       Herb                         3-Octanone
#> 3       Mold                         3-Octanone
#> 4       None                             Decane
#> 5       None                         Mesitylene
#> 6       None            1,2,4-trimethyl-benzene
#> 7     Citrus                         D-Limonene
#> 8       Mint                         D-Limonene
#> 9       None          beta-ethyl benzeneethanol
#> 10      None                1,4-diethyl benzene
#> 11      None                1,2-diethyl benzene
#> 12      None               1,3,8-p-Menthatriene
#> 13      None      (2-methyl-1-propenyl)-Benzene
#> 14      None                  1-Phenyl-1-butene
#> 15 Coriander                           Linalool
#> 16    Floral                           Linalool
#> 17  Lavender                           Linalool
#> 18     Lemon                           Linalool
#> 19      Rose                           Linalool
#> 20       Fat                            Nonanal
#> 21    Floral                            Nonanal
#> 22     Green                            Nonanal
#> 23     Lemon                            Nonanal
#> 24      None 5-nonyl-2-Thiophenecarboxylic acid
#> 25      None               Dichloroacetaldehyde
#> 26     Fruit                    Linalyl acetate
#> 27    Floral                       Beta-Ocimene
#> 
#> $FDA_SPL
#>                  FDA_SPL                           Chemical
#> 1                   None                         3-Octanone
#> 2                 DECANE                             Decane
#> 3             MESITYLENE                         Mesitylene
#> 4                   None            1,2,4-trimethyl-benzene
#> 5         LIMONENE, (+)-                         D-Limonene
#> 6                   None          beta-ethyl benzeneethanol
#> 7     1,4-DIETHYLBENZENE                1,4-diethyl benzene
#> 8     1,2-DIETHYLBENZENE                1,2-diethyl benzene
#> 9  P-MENTHA-1,3,8-TRIENE               1,3,8-p-Menthatriene
#> 10     ISOBUTENYLBENZENE      (2-methyl-1-propenyl)-Benzene
#> 11                  None                  1-Phenyl-1-butene
#> 12        LINALOOL, (±)-                           Linalool
#> 13               NONANAL                            Nonanal
#> 14                  None 5-nonyl-2-Thiophenecarboxylic acid
#> 15  DICHLOROACETALDEHYDE               Dichloroacetaldehyde
#> 16       LINALYL ACETATE                    Linalyl acetate
#> 17                  None                       Beta-Ocimene
#> 
#> $FMCS
#>                                             Chemical        MW        MF Rings
#> 1                                         3-octanone 128.21204    C8H16O  None
#> 2                                         3-octanone 128.21204    C8H16O  None
#> 3                                         3-octanone 128.21204    C8H16O  None
#> 4                                         3-octanone 128.21204    C8H16O  None
#> 5                                         3-octanone 128.21204    C8H16O  None
#> 6                                         3-octanone 128.21204    C8H16O  None
#> 7                                         3-octanone 128.21204    C8H16O  None
#> 8                                         3-octanone 128.21204    C8H16O  None
#> 9                                         3-octanone 128.21204    C8H16O  None
#> 10                                        3-octanone 128.21204    C8H16O  None
#> 11                                        3-octanone 128.21204    C8H16O  None
#> 12                                        3-octanone 128.21204    C8H16O  None
#> 13                                            decane 142.28168    C10H22  None
#> 14                                            decane 142.28168    C10H22  None
#> 15                                            decane 142.28168    C10H22  None
#> 16                                            decane 142.28168    C10H22  None
#> 17                                            decane 142.28168    C10H22  None
#> 18                                            decane 142.28168    C10H22  None
#> 19                                            decane 142.28168    C10H22  None
#> 20                                            decane 142.28168    C10H22  None
#> 21                                            decane 142.28168    C10H22  None
#> 22                                            decane 142.28168    C10H22  None
#> 23                                            decane 142.28168    C10H22  None
#> 24                                            decane 142.28168    C10H22  None
#> 25                            1,3,5-trimethylbenzene 120.19158     C9H12  None
#> 26                            1,3,5-trimethylbenzene 120.19158     C9H12  None
#> 27                            1,3,5-trimethylbenzene 120.19158     C9H12  None
#> 28                            1,3,5-trimethylbenzene 120.19158     C9H12  None
#> 29                            1,3,5-trimethylbenzene 120.19158     C9H12  None
#> 30                            1,3,5-trimethylbenzene 120.19158     C9H12  None
#> 31                            1,3,5-trimethylbenzene 120.19158     C9H12  None
#> 32                            1,3,5-trimethylbenzene 120.19158     C9H12  None
#> 33                            1,3,5-trimethylbenzene 120.19158     C9H12  None
#> 34                            1,3,5-trimethylbenzene 120.19158     C9H12  None
#> 35                            1,3,5-trimethylbenzene 120.19158     C9H12  None
#> 36                            1,3,5-trimethylbenzene 120.19158     C9H12  None
#> 37                            1,2,4-trimethylbenzene 120.19158     C9H12  None
#> 38                            1,2,4-trimethylbenzene 120.19158     C9H12  None
#> 39                            1,2,4-trimethylbenzene 120.19158     C9H12  None
#> 40                            1,2,4-trimethylbenzene 120.19158     C9H12  None
#> 41                            1,2,4-trimethylbenzene 120.19158     C9H12  None
#> 42                            1,2,4-trimethylbenzene 120.19158     C9H12  None
#> 43                            1,2,4-trimethylbenzene 120.19158     C9H12  None
#> 44                            1,2,4-trimethylbenzene 120.19158     C9H12  None
#> 45                            1,2,4-trimethylbenzene 120.19158     C9H12  None
#> 46                            1,2,4-trimethylbenzene 120.19158     C9H12  None
#> 47                            1,2,4-trimethylbenzene 120.19158     C9H12  None
#> 48                            1,2,4-trimethylbenzene 120.19158     C9H12  None
#> 49      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene 136.23404    C10H16  None
#> 50      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene 136.23404    C10H16  None
#> 51      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene 136.23404    C10H16  None
#> 52      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene 136.23404    C10H16  None
#> 53      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene 136.23404    C10H16  None
#> 54      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene 136.23404    C10H16  None
#> 55      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene 136.23404    C10H16  None
#> 56      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene 136.23404    C10H16  None
#> 57      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene 136.23404    C10H16  None
#> 58      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene 136.23404    C10H16  None
#> 59      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene 136.23404    C10H16  None
#> 60      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene 136.23404    C10H16  None
#> 61                                2-phenyl-1-butanol 150.21756   C10H14O  None
#> 62                                2-phenyl-1-butanol 150.21756   C10H14O  None
#> 63                                2-phenyl-1-butanol 150.21756   C10H14O  None
#> 64                                2-phenyl-1-butanol 150.21756   C10H14O  None
#> 65                                2-phenyl-1-butanol 150.21756   C10H14O  None
#> 66                                2-phenyl-1-butanol 150.21756   C10H14O  None
#> 67                                2-phenyl-1-butanol 150.21756   C10H14O  None
#> 68                                2-phenyl-1-butanol 150.21756   C10H14O  None
#> 69                                2-phenyl-1-butanol 150.21756   C10H14O  None
#> 70                                2-phenyl-1-butanol 150.21756   C10H14O  None
#> 71                                2-phenyl-1-butanol 150.21756   C10H14O  None
#> 72                                2-phenyl-1-butanol 150.21756   C10H14O  None
#> 73                                1,4-diethylbenzene 134.21816    C10H14  None
#> 74                                1,4-diethylbenzene 134.21816    C10H14  None
#> 75                                1,4-diethylbenzene 134.21816    C10H14  None
#> 76                                1,4-diethylbenzene 134.21816    C10H14  None
#> 77                                1,4-diethylbenzene 134.21816    C10H14  None
#> 78                                1,4-diethylbenzene 134.21816    C10H14  None
#> 79                                1,4-diethylbenzene 134.21816    C10H14  None
#> 80                                1,4-diethylbenzene 134.21816    C10H14  None
#> 81                                1,4-diethylbenzene 134.21816    C10H14  None
#> 82                                1,4-diethylbenzene 134.21816    C10H14  None
#> 83                                1,4-diethylbenzene 134.21816    C10H14  None
#> 84                                1,4-diethylbenzene 134.21816    C10H14  None
#> 85                                1,2-diethylbenzene 134.21816    C10H14  None
#> 86                                1,2-diethylbenzene 134.21816    C10H14  None
#> 87                                1,2-diethylbenzene 134.21816    C10H14  None
#> 88                                1,2-diethylbenzene 134.21816    C10H14  None
#> 89                                1,2-diethylbenzene 134.21816    C10H14  None
#> 90                                1,2-diethylbenzene 134.21816    C10H14  None
#> 91                                1,2-diethylbenzene 134.21816    C10H14  None
#> 92                                1,2-diethylbenzene 134.21816    C10H14  None
#> 93                                1,2-diethylbenzene 134.21816    C10H14  None
#> 94                                1,2-diethylbenzene 134.21816    C10H14  None
#> 95                                1,2-diethylbenzene 134.21816    C10H14  None
#> 96                                1,2-diethylbenzene 134.21816    C10H14  None
#> 97   1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene 134.21816    C10H14  None
#> 98   1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene 134.21816    C10H14  None
#> 99   1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene 134.21816    C10H14  None
#> 100  1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene 134.21816    C10H14  None
#> 101  1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene 134.21816    C10H14  None
#> 102  1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene 134.21816    C10H14  None
#> 103  1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene 134.21816    C10H14  None
#> 104  1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene 134.21816    C10H14  None
#> 105  1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene 134.21816    C10H14  None
#> 106  1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene 134.21816    C10H14  None
#> 107  1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene 134.21816    C10H14  None
#> 108  1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene 134.21816    C10H14  None
#> 109                       2-methylprop-1-enylbenzene 132.20228    C10H12  None
#> 110                       2-methylprop-1-enylbenzene 132.20228    C10H12  None
#> 111                       2-methylprop-1-enylbenzene 132.20228    C10H12  None
#> 112                       2-methylprop-1-enylbenzene 132.20228    C10H12  None
#> 113                       2-methylprop-1-enylbenzene 132.20228    C10H12  None
#> 114                       2-methylprop-1-enylbenzene 132.20228    C10H12  None
#> 115                       2-methylprop-1-enylbenzene 132.20228    C10H12  None
#> 116                       2-methylprop-1-enylbenzene 132.20228    C10H12  None
#> 117                       2-methylprop-1-enylbenzene 132.20228    C10H12  None
#> 118                       2-methylprop-1-enylbenzene 132.20228    C10H12  None
#> 119                       2-methylprop-1-enylbenzene 132.20228    C10H12  None
#> 120                       2-methylprop-1-enylbenzene 132.20228    C10H12  None
#> 121                                but-1-enylbenzene 132.20228    C10H12  None
#> 122                                but-1-enylbenzene 132.20228    C10H12  None
#> 123                                but-1-enylbenzene 132.20228    C10H12  None
#> 124                                but-1-enylbenzene 132.20228    C10H12  None
#> 125                                but-1-enylbenzene 132.20228    C10H12  None
#> 126                                but-1-enylbenzene 132.20228    C10H12  None
#> 127                                but-1-enylbenzene 132.20228    C10H12  None
#> 128                                but-1-enylbenzene 132.20228    C10H12  None
#> 129                                but-1-enylbenzene 132.20228    C10H12  None
#> 130                                but-1-enylbenzene 132.20228    C10H12  None
#> 131                                but-1-enylbenzene 132.20228    C10H12  None
#> 132                                but-1-enylbenzene 132.20228    C10H12  None
#> 133                   3,7-dimethyl-3-octa-1,6-dienol 154.24932   C10H18O  None
#> 134                   3,7-dimethyl-3-octa-1,6-dienol 154.24932   C10H18O  None
#> 135                   3,7-dimethyl-3-octa-1,6-dienol 154.24932   C10H18O  None
#> 136                   3,7-dimethyl-3-octa-1,6-dienol 154.24932   C10H18O  None
#> 137                   3,7-dimethyl-3-octa-1,6-dienol 154.24932   C10H18O  None
#> 138                   3,7-dimethyl-3-octa-1,6-dienol 154.24932   C10H18O  None
#> 139                   3,7-dimethyl-3-octa-1,6-dienol 154.24932   C10H18O  None
#> 140                   3,7-dimethyl-3-octa-1,6-dienol 154.24932   C10H18O  None
#> 141                   3,7-dimethyl-3-octa-1,6-dienol 154.24932   C10H18O  None
#> 142                   3,7-dimethyl-3-octa-1,6-dienol 154.24932   C10H18O  None
#> 143                   3,7-dimethyl-3-octa-1,6-dienol 154.24932   C10H18O  None
#> 144                   3,7-dimethyl-3-octa-1,6-dienol 154.24932   C10H18O  None
#> 145                                          nonanal 142.23862    C9H18O  None
#> 146                                          nonanal 142.23862    C9H18O  None
#> 147                                          nonanal 142.23862    C9H18O  None
#> 148                                          nonanal 142.23862    C9H18O  None
#> 149                                          nonanal 142.23862    C9H18O  None
#> 150                                          nonanal 142.23862    C9H18O  None
#> 151                                          nonanal 142.23862    C9H18O  None
#> 152                                          nonanal 142.23862    C9H18O  None
#> 153                                          nonanal 142.23862    C9H18O  None
#> 154                                          nonanal 142.23862    C9H18O  None
#> 155                                          nonanal 142.23862    C9H18O  None
#> 156                                          nonanal 142.23862    C9H18O  None
#> 157               5-nonyl-2-thiophenecarboxylic acid 254.38828 C14H22O2S  None
#> 158               5-nonyl-2-thiophenecarboxylic acid 254.38828 C14H22O2S  None
#> 159               5-nonyl-2-thiophenecarboxylic acid 254.38828 C14H22O2S  None
#> 160               5-nonyl-2-thiophenecarboxylic acid 254.38828 C14H22O2S  None
#> 161               5-nonyl-2-thiophenecarboxylic acid 254.38828 C14H22O2S  None
#> 162               5-nonyl-2-thiophenecarboxylic acid 254.38828 C14H22O2S  None
#> 163               5-nonyl-2-thiophenecarboxylic acid 254.38828 C14H22O2S  None
#> 164               5-nonyl-2-thiophenecarboxylic acid 254.38828 C14H22O2S  None
#> 165               5-nonyl-2-thiophenecarboxylic acid 254.38828 C14H22O2S  None
#> 166               5-nonyl-2-thiophenecarboxylic acid 254.38828 C14H22O2S  None
#> 167               5-nonyl-2-thiophenecarboxylic acid 254.38828 C14H22O2S  None
#> 168               5-nonyl-2-thiophenecarboxylic acid 254.38828 C14H22O2S  None
#> 169                         2,2-dichloroacetaldehyde 112.94268  C2H2Cl2O  None
#> 170                         2,2-dichloroacetaldehyde 112.94268  C2H2Cl2O  None
#> 171                         2,2-dichloroacetaldehyde 112.94268  C2H2Cl2O  None
#> 172                         2,2-dichloroacetaldehyde 112.94268  C2H2Cl2O  None
#> 173                         2,2-dichloroacetaldehyde 112.94268  C2H2Cl2O  None
#> 174                         2,2-dichloroacetaldehyde 112.94268  C2H2Cl2O  None
#> 175                         2,2-dichloroacetaldehyde 112.94268  C2H2Cl2O  None
#> 176                         2,2-dichloroacetaldehyde 112.94268  C2H2Cl2O  None
#> 177                         2,2-dichloroacetaldehyde 112.94268  C2H2Cl2O  None
#> 178                         2,2-dichloroacetaldehyde 112.94268  C2H2Cl2O  None
#> 179                         2,2-dichloroacetaldehyde 112.94268  C2H2Cl2O  None
#> 180                         2,2-dichloroacetaldehyde 112.94268  C2H2Cl2O  None
#> 181 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester   196.286  C12H20O2  None
#> 182 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester   196.286  C12H20O2  None
#> 183 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester   196.286  C12H20O2  None
#> 184 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester   196.286  C12H20O2  None
#> 185 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester   196.286  C12H20O2  None
#> 186 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester   196.286  C12H20O2  None
#> 187 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester   196.286  C12H20O2  None
#> 188 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester   196.286  C12H20O2  None
#> 189 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester   196.286  C12H20O2  None
#> 190 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester   196.286  C12H20O2  None
#> 191 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester   196.286  C12H20O2  None
#> 192 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester   196.286  C12H20O2  None
#> 193                    3,7-dimethylocta-1,3,6-triene 136.23404    C10H16  None
#> 194                    3,7-dimethylocta-1,3,6-triene 136.23404    C10H16  None
#> 195                    3,7-dimethylocta-1,3,6-triene 136.23404    C10H16  None
#> 196                    3,7-dimethylocta-1,3,6-triene 136.23404    C10H16  None
#> 197                    3,7-dimethylocta-1,3,6-triene 136.23404    C10H16  None
#> 198                    3,7-dimethylocta-1,3,6-triene 136.23404    C10H16  None
#> 199                    3,7-dimethylocta-1,3,6-triene 136.23404    C10H16  None
#> 200                    3,7-dimethylocta-1,3,6-triene 136.23404    C10H16  None
#> 201                    3,7-dimethylocta-1,3,6-triene 136.23404    C10H16  None
#> 202                    3,7-dimethylocta-1,3,6-triene 136.23404    C10H16  None
#> 203                    3,7-dimethylocta-1,3,6-triene 136.23404    C10H16  None
#> 204                    3,7-dimethylocta-1,3,6-triene 136.23404    C10H16  None
#>     Groups GroupCounts Atom AtomCounts NCharges
#> 1     CMP1           0    C          8        0
#> 2     CMP1           0    H         16        0
#> 3     CMP1           0    O          1        0
#> 4     CMP1           0    C          8        0
#> 5     CMP1           0    H         16        0
#> 6     CMP1           0    O          1        0
#> 7     CMP1           1    C          8        0
#> 8     CMP1           0    H         16        0
#> 9     CMP1           0    O          1        0
#> 10    CMP1           0    C          8        0
#> 11    CMP1           0    H         16        0
#> 12    CMP1           0    O          1        0
#> 13    CMP2           0    C         10        0
#> 14    CMP2           0    H         22        0
#> 15    CMP2           0    C         10        0
#> 16    CMP2           0    H         22        0
#> 17    CMP2           0    C         10        0
#> 18    CMP2           0    H         22        0
#> 19    CMP2           0    C         10        0
#> 20    CMP2           0    H         22        0
#> 21    CMP2           0    C         10        0
#> 22    CMP2           0    H         22        0
#> 23    CMP2           0    C         10        0
#> 24    CMP2           0    H         22        0
#> 25    CMP3           0    C          9        0
#> 26    CMP3           0    H         12        0
#> 27    CMP3           0    C          9        0
#> 28    CMP3           0    H         12        0
#> 29    CMP3           0    C          9        0
#> 30    CMP3           0    H         12        0
#> 31    CMP3           0    C          9        0
#> 32    CMP3           0    H         12        0
#> 33    CMP3           0    C          9        0
#> 34    CMP3           0    H         12        0
#> 35    CMP3           0    C          9        0
#> 36    CMP3           0    H         12        0
#> 37    CMP4           0    C          9        0
#> 38    CMP4           0    H         12        0
#> 39    CMP4           0    C          9        0
#> 40    CMP4           0    H         12        0
#> 41    CMP4           0    C          9        0
#> 42    CMP4           0    H         12        0
#> 43    CMP4           0    C          9        0
#> 44    CMP4           0    H         12        0
#> 45    CMP4           0    C          9        0
#> 46    CMP4           0    H         12        0
#> 47    CMP4           0    C          9        0
#> 48    CMP4           0    H         12        0
#> 49    CMP5           0    C         10        0
#> 50    CMP5           0    H         16        0
#> 51    CMP5           0    C         10        0
#> 52    CMP5           0    H         16        0
#> 53    CMP5           0    C         10        0
#> 54    CMP5           0    H         16        0
#> 55    CMP5           0    C         10        0
#> 56    CMP5           0    H         16        0
#> 57    CMP5           0    C         10        0
#> 58    CMP5           0    H         16        0
#> 59    CMP5           0    C         10        0
#> 60    CMP5           0    H         16        0
#> 61    CMP6           0    C         10        0
#> 62    CMP6           0    H         14        0
#> 63    CMP6           0    O          1        0
#> 64    CMP6           0    C         10        0
#> 65    CMP6           1    H         14        0
#> 66    CMP6           0    O          1        0
#> 67    CMP6           0    C         10        0
#> 68    CMP6           0    H         14        0
#> 69    CMP6           0    O          1        0
#> 70    CMP6           0    C         10        0
#> 71    CMP6           0    H         14        0
#> 72    CMP6           0    O          1        0
#> 73    CMP7           0    C         10        0
#> 74    CMP7           0    H         14        0
#> 75    CMP7           0    C         10        0
#> 76    CMP7           0    H         14        0
#> 77    CMP7           0    C         10        0
#> 78    CMP7           0    H         14        0
#> 79    CMP7           0    C         10        0
#> 80    CMP7           0    H         14        0
#> 81    CMP7           0    C         10        0
#> 82    CMP7           0    H         14        0
#> 83    CMP7           0    C         10        0
#> 84    CMP7           0    H         14        0
#> 85    CMP8           0    C         10        0
#> 86    CMP8           0    H         14        0
#> 87    CMP8           0    C         10        0
#> 88    CMP8           0    H         14        0
#> 89    CMP8           0    C         10        0
#> 90    CMP8           0    H         14        0
#> 91    CMP8           0    C         10        0
#> 92    CMP8           0    H         14        0
#> 93    CMP8           0    C         10        0
#> 94    CMP8           0    H         14        0
#> 95    CMP8           0    C         10        0
#> 96    CMP8           0    H         14        0
#> 97    CMP9           0    C         10        0
#> 98    CMP9           0    H         14        0
#> 99    CMP9           0    C         10        0
#> 100   CMP9           0    H         14        0
#> 101   CMP9           0    C         10        0
#> 102   CMP9           0    H         14        0
#> 103   CMP9           0    C         10        0
#> 104   CMP9           0    H         14        0
#> 105   CMP9           0    C         10        0
#> 106   CMP9           0    H         14        0
#> 107   CMP9           0    C         10        0
#> 108   CMP9           0    H         14        0
#> 109  CMP10           0    C         10        0
#> 110  CMP10           0    H         12        0
#> 111  CMP10           0    C         10        0
#> 112  CMP10           0    H         12        0
#> 113  CMP10           0    C         10        0
#> 114  CMP10           0    H         12        0
#> 115  CMP10           0    C         10        0
#> 116  CMP10           0    H         12        0
#> 117  CMP10           0    C         10        0
#> 118  CMP10           0    H         12        0
#> 119  CMP10           0    C         10        0
#> 120  CMP10           0    H         12        0
#> 121  CMP11           0    C         10        0
#> 122  CMP11           0    H         12        0
#> 123  CMP11           0    C         10        0
#> 124  CMP11           0    H         12        0
#> 125  CMP11           0    C         10        0
#> 126  CMP11           0    H         12        0
#> 127  CMP11           0    C         10        0
#> 128  CMP11           0    H         12        0
#> 129  CMP11           0    C         10        0
#> 130  CMP11           0    H         12        0
#> 131  CMP11           0    C         10        0
#> 132  CMP11           0    H         12        0
#> 133  CMP12           0    C         10        0
#> 134  CMP12           0    H         18        0
#> 135  CMP12           0    O          1        0
#> 136  CMP12           0    C         10        0
#> 137  CMP12           1    H         18        0
#> 138  CMP12           0    O          1        0
#> 139  CMP12           0    C         10        0
#> 140  CMP12           0    H         18        0
#> 141  CMP12           0    O          1        0
#> 142  CMP12           0    C         10        0
#> 143  CMP12           0    H         18        0
#> 144  CMP12           0    O          1        0
#> 145  CMP13           0    C          9        0
#> 146  CMP13           0    H         18        0
#> 147  CMP13           0    O          1        0
#> 148  CMP13           0    C          9        0
#> 149  CMP13           0    H         18        0
#> 150  CMP13           1    O          1        0
#> 151  CMP13           0    C          9        0
#> 152  CMP13           0    H         18        0
#> 153  CMP13           0    O          1        0
#> 154  CMP13           0    C          9        0
#> 155  CMP13           0    H         18        0
#> 156  CMP13           0    O          1        0
#> 157  CMP14           0    C         14        0
#> 158  CMP14           0    H         22        0
#> 159  CMP14           0    O          2        0
#> 160  CMP14           0    S          1        0
#> 161  CMP14           0    C         14        0
#> 162  CMP14           0    H         22        0
#> 163  CMP14           0    O          2        0
#> 164  CMP14           1    S          1        0
#> 165  CMP14           0    C         14        0
#> 166  CMP14           0    H         22        0
#> 167  CMP14           0    O          2        0
#> 168  CMP14           0    S          1        0
#> 169  CMP15           0    C          2        0
#> 170  CMP15           0   Cl          2        0
#> 171  CMP15           0    H          2        0
#> 172  CMP15           0    O          1        0
#> 173  CMP15           0    C          2        0
#> 174  CMP15           1   Cl          2        0
#> 175  CMP15           0    H          2        0
#> 176  CMP15           0    O          1        0
#> 177  CMP15           0    C          2        0
#> 178  CMP15           0   Cl          2        0
#> 179  CMP15           0    H          2        0
#> 180  CMP15           0    O          1        0
#> 181  CMP16           0    C         12        0
#> 182  CMP16           0    H         20        0
#> 183  CMP16           0    O          2        0
#> 184  CMP16           0    C         12        0
#> 185  CMP16           0    H         20        0
#> 186  CMP16           0    O          2        0
#> 187  CMP16           0    C         12        0
#> 188  CMP16           0    H         20        0
#> 189  CMP16           1    O          2        0
#> 190  CMP16           0    C         12        0
#> 191  CMP16           0    H         20        0
#> 192  CMP16           0    O          2        0
#> 193  CMP17           0    C         10        0
#> 194  CMP17           0    H         16        0
#> 195  CMP17           0    C         10        0
#> 196  CMP17           0    H         16        0
#> 197  CMP17           0    C         10        0
#> 198  CMP17           0    H         16        0
#> 199  CMP17           0    C         10        0
#> 200  CMP17           0    H         16        0
#> 201  CMP17           0    C         10        0
#> 202  CMP17           0    H         16        0
#> 203  CMP17           0    C         10        0
#> 204  CMP17           0    H         16        0
#> 
#> $FunctionalGroups
#>    Type.A Type.B Type.C Type.D Type.E
#> 1      No     No      ~    Yes    Yes
#> 2      No     No    Yes      ~    Yes
#> 3      No      ~     No     No     No
#> 4      No      ~     No     No     No
#> 5      No     No     No      ~      ~
#> 6      No      ~     No     No     No
#> 7      No      ~     No     No     No
#> 8      No      ~     No     No     No
#> 9      No     No     No      ~      ~
#> 10     No      ~     No     No     No
#> 11     No      ~     No     No     No
#> 12     No     No     No     No     No
#> 13     No     No    Yes      ~    Yes
#> 14     No     No    Yes      ~    Yes
#> 15     No     No     No     No     No
#> 16     No     No     No     No     No
#> 17     No     No     No     No     No
#>                                            Chemical
#> 1                                        3-octanone
#> 2                                            decane
#> 3                            1,3,5-trimethylbenzene
#> 4                            1,2,4-trimethylbenzene
#> 5      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene
#> 6                                2-phenyl-1-butanol
#> 7                                1,4-diethylbenzene
#> 8                                1,2-diethylbenzene
#> 9   1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene
#> 10                       2-methylprop-1-enylbenzene
#> 11                                but-1-enylbenzene
#> 12                   3,7-dimethyl-3-octa-1,6-dienol
#> 13                                          nonanal
#> 14               5-nonyl-2-thiophenecarboxylic acid
#> 15                         2,2-dichloroacetaldehyde
#> 16 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester
#> 17                    3,7-dimethylocta-1,3,6-triene
#> 
#> $BestChemMatch
#>    Type.A Type.B Type.C Type.D Type.E
#> 1      No     No     No   CMP7   CMP1
#> 2      No     No   CMP1     No   CMP1
#> 3      No     No     No     No     No
#> 4      No     No     No     No     No
#> 5      No     No     No     No     No
#> 6      No     No     No     No     No
#> 7      No     No     No     No     No
#> 8      No     No     No     No     No
#> 9      No     No     No     No     No
#> 10     No     No     No     No     No
#> 11     No     No     No     No     No
#> 12     No     No     No     No     No
#> 13     No     No   CMP1     No   CMP1
#> 14     No     No   CMP1     No   CMP1
#> 15     No     No     No     No     No
#> 16     No     No     No     No     No
#> 17     No     No     No     No     No
#>                                            Chemical
#> 1                                        3-octanone
#> 2                                            decane
#> 3                            1,3,5-trimethylbenzene
#> 4                            1,2,4-trimethylbenzene
#> 5      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene
#> 6                                2-phenyl-1-butanol
#> 7                                1,4-diethylbenzene
#> 8                                1,2-diethylbenzene
#> 9   1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene
#> 10                       2-methylprop-1-enylbenzene
#> 11                                but-1-enylbenzene
#> 12                   3,7-dimethyl-3-octa-1,6-dienol
#> 13                                          nonanal
#> 14               5-nonyl-2-thiophenecarboxylic acid
#> 15                         2,2-dichloroacetaldehyde
#> 16 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester
#> 17                    3,7-dimethylocta-1,3,6-triene
#>