categorate

Searches a vector of chemical names on PubChem for published information. Uses regular expressions to extract information from 6 data- bases: LOTUS the natural products occurrence database [LOTUS](https://lotus.naturalproducts.net/), Flavor and Extract Manufacturers Association [FEMA](https://www.femaflavor.org/), Kyoto Encyclopedia of Genes and Genomes [KEGG](https://www.genome.jp/kegg/), Food and Drug Administration [FDA/SPL](https://www.fda.gov/), and Reactive Groups from [PubChem](https://pubchem.ncbi.nlm.nih.gov/). Also downloads structural data in SDF format to summarize the atomic structures and common molecular groups of each chemical. Finally, the function downloads structural data for the chemicals within each group of the input library (`chemical_library`). Each chemical of interest (`compounds`) is tested against every group of chemicals for structural overlaps that exceed a threshold of 0.85.

categorate(compounds, chemical_library, input_format = "wide")

Arguments

compounds

A vector containing chemical names in IUPAC notation (preferred).

chemical_library

A .CSV file containing columns with chemical groups. Column names label the group while the rows contain chemicals that are described by the label.

input_format

Designates the data structure for input library (`chemical_library`). Default setting is "wide" can be changed to "long."

Value

List with 4 data frames: (1) information from database searches, (2) atomic and molecular sub-group structure summary, (3) strong (similarity > 0.95) or moderate (similarity > 0.85) matches with input library groups, and (4) top chemical from groups a chemical shared a strong match with.

Details

Provides a detailed overview of categorical and structural information for every chemical of interest. Functional matches are also generated by matching with groups from a chemical library.

Examples

compounds = c("3-Octanone","Decane","Mesitylene","1,2,4-trimethyl-benzene",
"D-Limonene","beta-ethyl benzeneethanol","1,4-diethyl benzene",
"1,2-diethyl benzene","1,3,8-p-Menthatriene","(2-methyl-1-propenyl)-Benzene",
"1-Phenyl-1-butene","Linalool","Nonanal","5-nonyl-2-Thiophenecarboxylic acid",
"Dichloroacetaldehyde","Linalyl acetate","Beta-Ocimene")
categorate(compounds, library_data, input_format = "wide")
#> Searching query chemicals for structural data, please be patient!
#> [1/17]-3-Octanone
#> [2/17]-Decane
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/15600/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/15600/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> [3/17]-Mesitylene
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/7947/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> [4/17]-1,2,4-trimethyl-benzene
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/7247/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> [5/17]-D-Limonene
#> [6/17]-beta-ethyl benzeneethanol
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/95542/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/95542/JSON?source=LOTUS+-+the+natural+products+occurrence+database': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/95542/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/95542/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/95542/JSON?source=Medical+Subject+Headings+(MeSH)': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/95542/JSON?source=FDA/SPL+Indexing+Data': HTTP status was '404 Not Found'
#> [7/17]-1,4-diethyl benzene
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/7734/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/7734/JSON?source=LOTUS+-+the+natural+products+occurrence+database': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/7734/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> [8/17]-1,2-diethyl benzene
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/8657/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/8657/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/8657/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> [9/17]-1,3,8-p-Menthatriene
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/176983/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/176983/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/176983/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/176983/JSON?source=Medical+Subject+Headings+(MeSH)': HTTP status was '404 Not Found'
#> [10/17]-(2-methyl-1-propenyl)-Benzene
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/13030/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/13030/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/13030/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/13030/JSON?source=Medical+Subject+Headings+(MeSH)': HTTP status was '404 Not Found'
#> [11/17]-1-Phenyl-1-butene
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/123088/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/123088/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/123088/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/123088/JSON?source=FDA/SPL+Indexing+Data': HTTP status was '404 Not Found'
#> [12/17]-Linalool
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/6549/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> [13/17]-Nonanal
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/31289/JSON?source=KEGG': HTTP status was '404 Not Found'
#> [14/17]-5-nonyl-2-Thiophenecarboxylic acid
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/587962/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/587962/JSON?source=LOTUS+-+the+natural+products+occurrence+database': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/587962/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/587962/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/587962/JSON?source=Medical+Subject+Headings+(MeSH)': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/587962/JSON?source=FDA/SPL+Indexing+Data': HTTP status was '404 Not Found'
#> [15/17]-Dichloroacetaldehyde
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/6576/JSON?source=LOTUS+-+the+natural+products+occurrence+database': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/6576/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> [16/17]-Linalyl acetate
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/8294/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/8294/JSON?source=KEGG': HTTP status was '404 Not Found'
#> [17/17]-Beta-Ocimene
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/18756/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/18756/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/18756/JSON?source=FDA/SPL+Indexing+Data': HTTP status was '404 Not Found'
#> Setting up your input library for Tanimoto matching...
#> [1/5]-Type.A
#> [1/4]-2-Aminothiazole
#> [2/4]-3,4,5,6-Tetrachlorocyclohexene
#> [3/4]-N-methyl-1,3,5-triazin-2-amine
#> [4/4]-2-Methyloctahydro-2-azacyclopropa[cd]pentalene
#> [2/5]-Type.B
#> [1/6]-o-Cresol
#> [2/6]-Salicylic Acid
#> [3/6]-Guaiacol
#> [4/6]-Aspirin
#> [5/6]-Salicyl alcohol
#> [6/6]-4-Methylsalicylic acid
#> [3/5]-Type.C
#> [1/3]-Nonane
#> [2/3]-Dodecane
#> [3/3]-Tridecane
#> [4/5]-Type.D
#> [1/7]-Butyl methacrylate
#> [2/7]-Isobutyl hexanoate
#> [3/7]-Ethyl heptanoate
#> [4/7]-Methyl hexanoate
#> [5/7]-Methyl heptanoate
#> [6/7]-Dihexyl adipate
#> [7/7]-2-Heptanone
#> [5/5]-Type.E
#> [1/7]-Octane
#> [2/7]-Octanoic acid
#> [3/7]-1-Octanol
#> [4/7]-Hexadecanal
#> [5/7]-Decanal
#> [6/7]-Undecanal
#> [7/7]-Hexyl acetate
#> Warning: URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug/compound/cid/335,338,460,2244,5146,5788/SDF': Timeout of 60 seconds was reached
#> Error in file(con, "r"): cannot open the connection to 'https://pubchem.ncbi.nlm.nih.gov/rest/pug/compound/cid/335,338,460,2244,5146,5788/SDF'