categorate

Searches a vector of chemical names on PubChem for published information. Uses regular expressions to extract information from 6 data- bases: LOTUS the natural products occurrence database [LOTUS](https://lotus.naturalproducts.net/), Flavor and Extract Manufacturers Association [FEMA](https://www.femaflavor.org/), Kyoto Encyclopedia of Genes and Genomes [KEGG](https://www.genome.jp/kegg/), Food and Drug Administration [FDA/SPL](https://www.fda.gov/), and Reactive Groups from [PubChem](https://pubchem.ncbi.nlm.nih.gov/). Also downloads structural data in SDF format to summarize the atomic structures and common molecular groups of each chemical. Finally, the function downloads structural data for the chemicals within each group of the input library (`chemical_library`). Each chemical of interest (`compounds`) is tested against every group of chemicals for structural overlaps that exceed a threshold of 0.85.

categorate(compounds, chemical_library, input_format = "wide")

Arguments

compounds

A vector containing chemical names in IUPAC notation (preferred).

chemical_library

A .CSV file containing columns with chemical groups. Column names label the group while the rows contain chemicals that are described by the label.

input_format

Designates the data structure for input library (`chemical_library`). Default setting is "wide" can be changed to "long."

Value

List with 4 data frames: (1) information from database searches, (2) atomic and molecular sub-group structure summary, (3) strong (similarity > 0.95) or moderate (similarity > 0.85) matches with input library groups, and (4) top chemical from groups a chemical shared a strong match with.

Details

Provides a detailed overview of categorical and structural information for every chemical of interest. Functional matches are also generated by matching with groups from a chemical library.

Examples

compounds = c("3-Octanone","Decane","Mesitylene","1,2,4-trimethyl-benzene",
"D-Limonene","beta-ethyl benzeneethanol","1,4-diethyl benzene",
"1,2-diethyl benzene","1,3,8-p-Menthatriene","(2-methyl-1-propenyl)-Benzene",
"1-Phenyl-1-butene","Linalool","Nonanal","5-nonyl-2-Thiophenecarboxylic acid",
"Dichloroacetaldehyde","Linalyl acetate","Beta-Ocimene")
categorate(compounds, library_data, input_format = "wide")
#> Searching query chemicals for structural data, please be patient!
#> [1/17]-3-Octanone
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/246728/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: number of columns of result, 3, is not a multiple of vector length 2 of arg 1
#> Warning: number of columns of result, 3, is not a multiple of vector length 2 of arg 2
#> [2/17]-Decane
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/15600/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/15600/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> Warning: number of columns of result, 3, is not a multiple of vector length 2 of arg 2
#> [3/17]-Mesitylene
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/7947/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/7947/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> [4/17]-1,2,4-trimethyl-benzene
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/7247/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/7247/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> [5/17]-D-Limonene
#> [6/17]-beta-ethyl benzeneethanol
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/95542/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/95542/JSON?source=LOTUS+-+the+natural+products+occurrence+database': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/95542/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/95542/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/95542/JSON?source=Medical+Subject+Headings+(MeSH)': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/95542/JSON?source=FDA/SPL+Indexing+Data': HTTP status was '404 Not Found'
#> [7/17]-1,4-diethyl benzene
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/7734/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/7734/JSON?source=LOTUS+-+the+natural+products+occurrence+database': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/7734/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/7734/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> [8/17]-1,2-diethyl benzene
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/8657/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/8657/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/8657/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> [9/17]-1,3,8-p-Menthatriene
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/176983/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/176983/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/176983/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/176983/JSON?source=Medical+Subject+Headings+(MeSH)': HTTP status was '404 Not Found'
#> [10/17]-(2-methyl-1-propenyl)-Benzene
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/13030/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/13030/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/13030/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/13030/JSON?source=Medical+Subject+Headings+(MeSH)': HTTP status was '404 Not Found'
#> [11/17]-1-Phenyl-1-butene
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/5370622/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/5370622/JSON?source=LOTUS+-+the+natural+products+occurrence+database': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/5370622/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/5370622/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/5370622/JSON?source=FDA/SPL+Indexing+Data': HTTP status was '404 Not Found'
#> [12/17]-Linalool
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/6549/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: number of columns of result, 5, is not a multiple of vector length 2 of arg 2
#> Warning: number of columns of result, 5, is not a multiple of vector length 2 of arg 3
#> [13/17]-Nonanal
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/31289/JSON?source=KEGG': HTTP status was '404 Not Found'
#> [14/17]-5-nonyl-2-Thiophenecarboxylic acid
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/587962/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/587962/JSON?source=LOTUS+-+the+natural+products+occurrence+database': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/587962/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/587962/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/587962/JSON?source=Medical+Subject+Headings+(MeSH)': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/587962/JSON?source=FDA/SPL+Indexing+Data': HTTP status was '404 Not Found'
#> [15/17]-Dichloroacetaldehyde
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/6576/JSON?source=LOTUS+-+the+natural+products+occurrence+database': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/6576/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/6576/JSON?source=Flavor+and+Extract+Manufacturers+Association+(FEMA)': HTTP status was '404 Not Found'
#> [16/17]-Linalyl acetate
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/8294/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/8294/JSON?source=KEGG': HTTP status was '404 Not Found'
#> [17/17]-Beta-Ocimene
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/18756/JSON?heading=Reactive+Group': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/18756/JSON?source=KEGG': HTTP status was '404 Not Found'
#> Warning: cannot open URL 'https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data/compound/18756/JSON?source=FDA/SPL+Indexing+Data': HTTP status was '404 Not Found'
#> Setting up your input library for Tanimoto matching...
#> [1/5]-Type.A
#> [1/4]-2-Aminothiazole
#> [2/4]-3,4,5,6-Tetrachlorocyclohexene
#> [3/4]-N-methyl-1,3,5-triazin-2-amine
#> [4/4]-2-Methyloctahydro-2-azacyclopropa[cd]pentalene
#> [2/5]-Type.B
#> [1/6]-o-Cresol
#> [2/6]-Salicylic Acid
#> [3/6]-Guaiacol
#> [4/6]-Aspirin
#> [5/6]-Salicyl alcohol
#> [6/6]-4-Methylsalicylic acid
#> [3/5]-Type.C
#> [1/3]-Nonane
#> [2/3]-Dodecane
#> [3/3]-Tridecane
#> [4/5]-Type.D
#> [1/7]-Butyl methacrylate
#> [2/7]-Isobutyl hexanoate
#> [3/7]-Ethyl heptanoate
#> [4/7]-Methyl hexanoate
#> [5/7]-Methyl heptanoate
#> [6/7]-Dihexyl adipate
#> [7/7]-2-Heptanone
#> [5/5]-Type.E
#> [1/7]-Octane
#> [2/7]-Octanoic acid
#> [3/7]-1-Octanol
#> [4/7]-Hexadecanal
#> [5/7]-Decanal
#> [6/7]-Undecanal
#> [7/7]-Hexyl acetate
#> [1] "No duplicates detected!"
#> $Databases
#>                           reactives_df  LOTUS_df                       KEGG_df
#> 1                              Ketones Q18349104                          None
#> 2                    ETHYL AMYL KETONE   biochem                          None
#> 3                              Ketones Q18349104                          None
#> 4                               Decane   Q150717                          None
#> 5    Hydrocarbons, Aliphatic Saturated   biochem                          None
#> 6                             N-DECANE   Q150717                          None
#> 7               Hydrocarbons, Aromatic   Q425161                          None
#> 8               Hydrocarbons, Aromatic   biochem                          None
#> 9               Hydrocarbons, Aromatic   Q376994                          None
#> 10              Hydrocarbons, Aromatic   biochem                          None
#> 11 Hydrocarbons, Aliphatic Unsaturated Q27888324                   KEGG: Lipid
#> 12 Hydrocarbons, Aliphatic Unsaturated   biochem KEGG: Phytochemical Compounds
#> 13                                None      None                          None
#> 14                                None      None                          None
#> 15                                None  Q2429316                          None
#> 16                                None   biochem                          None
#> 17                                None Q27161428                          None
#> 18                                None   biochem                          None
#> 19                                None Q81989319                          None
#> 20                                None   biochem                          None
#> 21                                None      None                          None
#> 22                                None   Q410932                   KEGG: Lipid
#> 23                                None   biochem KEGG: Phytochemical Compounds
#> 24                                None   Q410932                   KEGG: Lipid
#> 25                                None   biochem KEGG: Phytochemical Compounds
#> 26                                None   Q410932                   KEGG: Lipid
#> 27                           Aldehydes   Q419668                          None
#> 28                           Aldehydes   biochem                          None
#> 29                           Aldehydes   Q419668                          None
#> 30                           Aldehydes   biochem                          None
#> 31                                None      None                          None
#> 32                Dichloroacetaldehyde      None                          None
#> 33                           Aldehydes      None                          None
#> 34       Halogenated Organic Compounds      None                          None
#> 35             Polymerizable Compounds      None                          None
#> 36                                None   Q188314                          None
#> 37                                None   biochem                          None
#> 38                                None Q27108635                          None
#> 39                                None   biochem                          None
#>      FEMA_df            FDA_SPL_df                           Chemical
#> 1     Butter            3-OCTANONE                         3-Octanone
#> 2       Herb            3-OCTANONE                         3-Octanone
#> 3       Mold            3-OCTANONE                         3-Octanone
#> 4       None                DECANE                             Decane
#> 5       None                DECANE                             Decane
#> 6       None                DECANE                             Decane
#> 7       None            MESITYLENE                         Mesitylene
#> 8       None            MESITYLENE                         Mesitylene
#> 9       None          PSEUDOCUMENE            1,2,4-trimethyl-benzene
#> 10      None          PSEUDOCUMENE            1,2,4-trimethyl-benzene
#> 11    Citrus        LIMONENE, (+)-                         D-Limonene
#> 12      Mint        LIMONENE, (+)-                         D-Limonene
#> 13      None                  None          beta-ethyl benzeneethanol
#> 14      None    1,4-DIETHYLBENZENE                1,4-diethyl benzene
#> 15      None    1,2-DIETHYLBENZENE                1,2-diethyl benzene
#> 16      None    1,2-DIETHYLBENZENE                1,2-diethyl benzene
#> 17      None P-MENTHA-1,3,8-TRIENE               1,3,8-p-Menthatriene
#> 18      None P-MENTHA-1,3,8-TRIENE               1,3,8-p-Menthatriene
#> 19      None     ISOBUTENYLBENZENE      (2-methyl-1-propenyl)-Benzene
#> 20      None     ISOBUTENYLBENZENE      (2-methyl-1-propenyl)-Benzene
#> 21      None                  None                  1-Phenyl-1-butene
#> 22 Coriander        LINALOOL, (±)-                           Linalool
#> 23    Floral        LINALOOL, (±)-                           Linalool
#> 24  Lavender        LINALOOL, (±)-                           Linalool
#> 25     Lemon        LINALOOL, (±)-                           Linalool
#> 26      Rose        LINALOOL, (±)-                           Linalool
#> 27       Fat               NONANAL                            Nonanal
#> 28    Floral               NONANAL                            Nonanal
#> 29     Green               NONANAL                            Nonanal
#> 30     Lemon               NONANAL                            Nonanal
#> 31      None                  None 5-nonyl-2-Thiophenecarboxylic acid
#> 32      None  DICHLOROACETALDEHYDE               Dichloroacetaldehyde
#> 33      None  DICHLOROACETALDEHYDE               Dichloroacetaldehyde
#> 34      None  DICHLOROACETALDEHYDE               Dichloroacetaldehyde
#> 35      None  DICHLOROACETALDEHYDE               Dichloroacetaldehyde
#> 36     Fruit       LINALYL ACETATE                    Linalyl acetate
#> 37     Fruit       LINALYL ACETATE                    Linalyl acetate
#> 38    Floral                  None                       Beta-Ocimene
#> 39    Floral                  None                       Beta-Ocimene
#> 
#> $FMCS
#>                                             Chemical        MW        MF Rings
#> 1                                         3-octanone 128.21204    C8H16O  None
#> 2                                         3-octanone 128.21204    C8H16O  None
#> 3                                         3-octanone 128.21204    C8H16O  None
#> 4                                         3-octanone 128.21204    C8H16O  None
#> 5                                         3-octanone 128.21204    C8H16O  None
#> 6                                         3-octanone 128.21204    C8H16O  None
#> 7                                         3-octanone 128.21204    C8H16O  None
#> 8                                         3-octanone 128.21204    C8H16O  None
#> 9                                         3-octanone 128.21204    C8H16O  None
#> 10                                        3-octanone 128.21204    C8H16O  None
#> 11                                        3-octanone 128.21204    C8H16O  None
#> 12                                        3-octanone 128.21204    C8H16O  None
#> 13                                            decane 142.28168    C10H22  None
#> 14                                            decane 142.28168    C10H22  None
#> 15                                            decane 142.28168    C10H22  None
#> 16                                            decane 142.28168    C10H22  None
#> 17                                            decane 142.28168    C10H22  None
#> 18                                            decane 142.28168    C10H22  None
#> 19                                            decane 142.28168    C10H22  None
#> 20                                            decane 142.28168    C10H22  None
#> 21                                            decane 142.28168    C10H22  None
#> 22                                            decane 142.28168    C10H22  None
#> 23                                            decane 142.28168    C10H22  None
#> 24                                            decane 142.28168    C10H22  None
#> 25                            1,3,5-trimethylbenzene 120.19158     C9H12  None
#> 26                            1,3,5-trimethylbenzene 120.19158     C9H12  None
#> 27                            1,3,5-trimethylbenzene 120.19158     C9H12  None
#> 28                            1,3,5-trimethylbenzene 120.19158     C9H12  None
#> 29                            1,3,5-trimethylbenzene 120.19158     C9H12  None
#> 30                            1,3,5-trimethylbenzene 120.19158     C9H12  None
#> 31                            1,3,5-trimethylbenzene 120.19158     C9H12  None
#> 32                            1,3,5-trimethylbenzene 120.19158     C9H12  None
#> 33                            1,3,5-trimethylbenzene 120.19158     C9H12  None
#> 34                            1,3,5-trimethylbenzene 120.19158     C9H12  None
#> 35                            1,3,5-trimethylbenzene 120.19158     C9H12  None
#> 36                            1,3,5-trimethylbenzene 120.19158     C9H12  None
#> 37                            1,2,4-trimethylbenzene 120.19158     C9H12  None
#> 38                            1,2,4-trimethylbenzene 120.19158     C9H12  None
#> 39                            1,2,4-trimethylbenzene 120.19158     C9H12  None
#> 40                            1,2,4-trimethylbenzene 120.19158     C9H12  None
#> 41                            1,2,4-trimethylbenzene 120.19158     C9H12  None
#> 42                            1,2,4-trimethylbenzene 120.19158     C9H12  None
#> 43                            1,2,4-trimethylbenzene 120.19158     C9H12  None
#> 44                            1,2,4-trimethylbenzene 120.19158     C9H12  None
#> 45                            1,2,4-trimethylbenzene 120.19158     C9H12  None
#> 46                            1,2,4-trimethylbenzene 120.19158     C9H12  None
#> 47                            1,2,4-trimethylbenzene 120.19158     C9H12  None
#> 48                            1,2,4-trimethylbenzene 120.19158     C9H12  None
#> 49      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene 136.23404    C10H16  None
#> 50      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene 136.23404    C10H16  None
#> 51      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene 136.23404    C10H16  None
#> 52      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene 136.23404    C10H16  None
#> 53      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene 136.23404    C10H16  None
#> 54      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene 136.23404    C10H16  None
#> 55      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene 136.23404    C10H16  None
#> 56      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene 136.23404    C10H16  None
#> 57      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene 136.23404    C10H16  None
#> 58      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene 136.23404    C10H16  None
#> 59      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene 136.23404    C10H16  None
#> 60      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene 136.23404    C10H16  None
#> 61                                2-phenyl-1-butanol 150.21756   C10H14O  None
#> 62                                2-phenyl-1-butanol 150.21756   C10H14O  None
#> 63                                2-phenyl-1-butanol 150.21756   C10H14O  None
#> 64                                2-phenyl-1-butanol 150.21756   C10H14O  None
#> 65                                2-phenyl-1-butanol 150.21756   C10H14O  None
#> 66                                2-phenyl-1-butanol 150.21756   C10H14O  None
#> 67                                2-phenyl-1-butanol 150.21756   C10H14O  None
#> 68                                2-phenyl-1-butanol 150.21756   C10H14O  None
#> 69                                2-phenyl-1-butanol 150.21756   C10H14O  None
#> 70                                2-phenyl-1-butanol 150.21756   C10H14O  None
#> 71                                2-phenyl-1-butanol 150.21756   C10H14O  None
#> 72                                2-phenyl-1-butanol 150.21756   C10H14O  None
#> 73                                1,4-diethylbenzene 134.21816    C10H14  None
#> 74                                1,4-diethylbenzene 134.21816    C10H14  None
#> 75                                1,4-diethylbenzene 134.21816    C10H14  None
#> 76                                1,4-diethylbenzene 134.21816    C10H14  None
#> 77                                1,4-diethylbenzene 134.21816    C10H14  None
#> 78                                1,4-diethylbenzene 134.21816    C10H14  None
#> 79                                1,4-diethylbenzene 134.21816    C10H14  None
#> 80                                1,4-diethylbenzene 134.21816    C10H14  None
#> 81                                1,4-diethylbenzene 134.21816    C10H14  None
#> 82                                1,4-diethylbenzene 134.21816    C10H14  None
#> 83                                1,4-diethylbenzene 134.21816    C10H14  None
#> 84                                1,4-diethylbenzene 134.21816    C10H14  None
#> 85                                1,2-diethylbenzene 134.21816    C10H14  None
#> 86                                1,2-diethylbenzene 134.21816    C10H14  None
#> 87                                1,2-diethylbenzene 134.21816    C10H14  None
#> 88                                1,2-diethylbenzene 134.21816    C10H14  None
#> 89                                1,2-diethylbenzene 134.21816    C10H14  None
#> 90                                1,2-diethylbenzene 134.21816    C10H14  None
#> 91                                1,2-diethylbenzene 134.21816    C10H14  None
#> 92                                1,2-diethylbenzene 134.21816    C10H14  None
#> 93                                1,2-diethylbenzene 134.21816    C10H14  None
#> 94                                1,2-diethylbenzene 134.21816    C10H14  None
#> 95                                1,2-diethylbenzene 134.21816    C10H14  None
#> 96                                1,2-diethylbenzene 134.21816    C10H14  None
#> 97   1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene 134.21816    C10H14  None
#> 98   1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene 134.21816    C10H14  None
#> 99   1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene 134.21816    C10H14  None
#> 100  1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene 134.21816    C10H14  None
#> 101  1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene 134.21816    C10H14  None
#> 102  1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene 134.21816    C10H14  None
#> 103  1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene 134.21816    C10H14  None
#> 104  1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene 134.21816    C10H14  None
#> 105  1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene 134.21816    C10H14  None
#> 106  1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene 134.21816    C10H14  None
#> 107  1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene 134.21816    C10H14  None
#> 108  1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene 134.21816    C10H14  None
#> 109                       2-methylprop-1-enylbenzene 132.20228    C10H12  None
#> 110                       2-methylprop-1-enylbenzene 132.20228    C10H12  None
#> 111                       2-methylprop-1-enylbenzene 132.20228    C10H12  None
#> 112                       2-methylprop-1-enylbenzene 132.20228    C10H12  None
#> 113                       2-methylprop-1-enylbenzene 132.20228    C10H12  None
#> 114                       2-methylprop-1-enylbenzene 132.20228    C10H12  None
#> 115                       2-methylprop-1-enylbenzene 132.20228    C10H12  None
#> 116                       2-methylprop-1-enylbenzene 132.20228    C10H12  None
#> 117                       2-methylprop-1-enylbenzene 132.20228    C10H12  None
#> 118                       2-methylprop-1-enylbenzene 132.20228    C10H12  None
#> 119                       2-methylprop-1-enylbenzene 132.20228    C10H12  None
#> 120                       2-methylprop-1-enylbenzene 132.20228    C10H12  None
#> 121                          [(E)-but-1-enyl]benzene 132.20228    C10H12  None
#> 122                          [(E)-but-1-enyl]benzene 132.20228    C10H12  None
#> 123                          [(E)-but-1-enyl]benzene 132.20228    C10H12  None
#> 124                          [(E)-but-1-enyl]benzene 132.20228    C10H12  None
#> 125                          [(E)-but-1-enyl]benzene 132.20228    C10H12  None
#> 126                          [(E)-but-1-enyl]benzene 132.20228    C10H12  None
#> 127                          [(E)-but-1-enyl]benzene 132.20228    C10H12  None
#> 128                          [(E)-but-1-enyl]benzene 132.20228    C10H12  None
#> 129                          [(E)-but-1-enyl]benzene 132.20228    C10H12  None
#> 130                          [(E)-but-1-enyl]benzene 132.20228    C10H12  None
#> 131                          [(E)-but-1-enyl]benzene 132.20228    C10H12  None
#> 132                          [(E)-but-1-enyl]benzene 132.20228    C10H12  None
#> 133                   3,7-dimethyl-3-octa-1,6-dienol 154.24932   C10H18O  None
#> 134                   3,7-dimethyl-3-octa-1,6-dienol 154.24932   C10H18O  None
#> 135                   3,7-dimethyl-3-octa-1,6-dienol 154.24932   C10H18O  None
#> 136                   3,7-dimethyl-3-octa-1,6-dienol 154.24932   C10H18O  None
#> 137                   3,7-dimethyl-3-octa-1,6-dienol 154.24932   C10H18O  None
#> 138                   3,7-dimethyl-3-octa-1,6-dienol 154.24932   C10H18O  None
#> 139                   3,7-dimethyl-3-octa-1,6-dienol 154.24932   C10H18O  None
#> 140                   3,7-dimethyl-3-octa-1,6-dienol 154.24932   C10H18O  None
#> 141                   3,7-dimethyl-3-octa-1,6-dienol 154.24932   C10H18O  None
#> 142                   3,7-dimethyl-3-octa-1,6-dienol 154.24932   C10H18O  None
#> 143                   3,7-dimethyl-3-octa-1,6-dienol 154.24932   C10H18O  None
#> 144                   3,7-dimethyl-3-octa-1,6-dienol 154.24932   C10H18O  None
#> 145                                          nonanal 142.23862    C9H18O  None
#> 146                                          nonanal 142.23862    C9H18O  None
#> 147                                          nonanal 142.23862    C9H18O  None
#> 148                                          nonanal 142.23862    C9H18O  None
#> 149                                          nonanal 142.23862    C9H18O  None
#> 150                                          nonanal 142.23862    C9H18O  None
#> 151                                          nonanal 142.23862    C9H18O  None
#> 152                                          nonanal 142.23862    C9H18O  None
#> 153                                          nonanal 142.23862    C9H18O  None
#> 154                                          nonanal 142.23862    C9H18O  None
#> 155                                          nonanal 142.23862    C9H18O  None
#> 156                                          nonanal 142.23862    C9H18O  None
#> 157               5-nonyl-2-thiophenecarboxylic acid 254.38828 C14H22O2S  None
#> 158               5-nonyl-2-thiophenecarboxylic acid 254.38828 C14H22O2S  None
#> 159               5-nonyl-2-thiophenecarboxylic acid 254.38828 C14H22O2S  None
#> 160               5-nonyl-2-thiophenecarboxylic acid 254.38828 C14H22O2S  None
#> 161               5-nonyl-2-thiophenecarboxylic acid 254.38828 C14H22O2S  None
#> 162               5-nonyl-2-thiophenecarboxylic acid 254.38828 C14H22O2S  None
#> 163               5-nonyl-2-thiophenecarboxylic acid 254.38828 C14H22O2S  None
#> 164               5-nonyl-2-thiophenecarboxylic acid 254.38828 C14H22O2S  None
#> 165               5-nonyl-2-thiophenecarboxylic acid 254.38828 C14H22O2S  None
#> 166               5-nonyl-2-thiophenecarboxylic acid 254.38828 C14H22O2S  None
#> 167               5-nonyl-2-thiophenecarboxylic acid 254.38828 C14H22O2S  None
#> 168               5-nonyl-2-thiophenecarboxylic acid 254.38828 C14H22O2S  None
#> 169                         2,2-dichloroacetaldehyde 112.94268  C2H2Cl2O  None
#> 170                         2,2-dichloroacetaldehyde 112.94268  C2H2Cl2O  None
#> 171                         2,2-dichloroacetaldehyde 112.94268  C2H2Cl2O  None
#> 172                         2,2-dichloroacetaldehyde 112.94268  C2H2Cl2O  None
#> 173                         2,2-dichloroacetaldehyde 112.94268  C2H2Cl2O  None
#> 174                         2,2-dichloroacetaldehyde 112.94268  C2H2Cl2O  None
#> 175                         2,2-dichloroacetaldehyde 112.94268  C2H2Cl2O  None
#> 176                         2,2-dichloroacetaldehyde 112.94268  C2H2Cl2O  None
#> 177                         2,2-dichloroacetaldehyde 112.94268  C2H2Cl2O  None
#> 178                         2,2-dichloroacetaldehyde 112.94268  C2H2Cl2O  None
#> 179                         2,2-dichloroacetaldehyde 112.94268  C2H2Cl2O  None
#> 180                         2,2-dichloroacetaldehyde 112.94268  C2H2Cl2O  None
#> 181 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester   196.286  C12H20O2  None
#> 182 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester   196.286  C12H20O2  None
#> 183 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester   196.286  C12H20O2  None
#> 184 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester   196.286  C12H20O2  None
#> 185 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester   196.286  C12H20O2  None
#> 186 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester   196.286  C12H20O2  None
#> 187 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester   196.286  C12H20O2  None
#> 188 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester   196.286  C12H20O2  None
#> 189 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester   196.286  C12H20O2  None
#> 190 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester   196.286  C12H20O2  None
#> 191 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester   196.286  C12H20O2  None
#> 192 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester   196.286  C12H20O2  None
#> 193                    3,7-dimethylocta-1,3,6-triene 136.23404    C10H16  None
#> 194                    3,7-dimethylocta-1,3,6-triene 136.23404    C10H16  None
#> 195                    3,7-dimethylocta-1,3,6-triene 136.23404    C10H16  None
#> 196                    3,7-dimethylocta-1,3,6-triene 136.23404    C10H16  None
#> 197                    3,7-dimethylocta-1,3,6-triene 136.23404    C10H16  None
#> 198                    3,7-dimethylocta-1,3,6-triene 136.23404    C10H16  None
#> 199                    3,7-dimethylocta-1,3,6-triene 136.23404    C10H16  None
#> 200                    3,7-dimethylocta-1,3,6-triene 136.23404    C10H16  None
#> 201                    3,7-dimethylocta-1,3,6-triene 136.23404    C10H16  None
#> 202                    3,7-dimethylocta-1,3,6-triene 136.23404    C10H16  None
#> 203                    3,7-dimethylocta-1,3,6-triene 136.23404    C10H16  None
#> 204                    3,7-dimethylocta-1,3,6-triene 136.23404    C10H16  None
#>     Groups GroupCounts Atom AtomCounts NCharges
#> 1     RNH2           0    C          8        0
#> 2     R2NH           0    H         16        0
#> 3      R3N           0    O          1        0
#> 4    ROPO3           0    C          8        0
#> 5      ROH           0    H         16        0
#> 6     RCHO           0    O          1        0
#> 7     RCOR           1    C          8        0
#> 8    RCOOH           0    H         16        0
#> 9    RCOOR           0    O          1        0
#> 10     ROR           0    C          8        0
#> 11    RCCH           0    H         16        0
#> 12     RCN           0    O          1        0
#> 13    RNH2           0    C         10        0
#> 14    R2NH           0    H         22        0
#> 15     R3N           0    C         10        0
#> 16   ROPO3           0    H         22        0
#> 17     ROH           0    C         10        0
#> 18    RCHO           0    H         22        0
#> 19    RCOR           0    C         10        0
#> 20   RCOOH           0    H         22        0
#> 21   RCOOR           0    C         10        0
#> 22     ROR           0    H         22        0
#> 23    RCCH           0    C         10        0
#> 24     RCN           0    H         22        0
#> 25    RNH2           0    C          9        0
#> 26    R2NH           0    H         12        0
#> 27     R3N           0    C          9        0
#> 28   ROPO3           0    H         12        0
#> 29     ROH           0    C          9        0
#> 30    RCHO           0    H         12        0
#> 31    RCOR           0    C          9        0
#> 32   RCOOH           0    H         12        0
#> 33   RCOOR           0    C          9        0
#> 34     ROR           0    H         12        0
#> 35    RCCH           0    C          9        0
#> 36     RCN           0    H         12        0
#> 37    RNH2           0    C          9        0
#> 38    R2NH           0    H         12        0
#> 39     R3N           0    C          9        0
#> 40   ROPO3           0    H         12        0
#> 41     ROH           0    C          9        0
#> 42    RCHO           0    H         12        0
#> 43    RCOR           0    C          9        0
#> 44   RCOOH           0    H         12        0
#> 45   RCOOR           0    C          9        0
#> 46     ROR           0    H         12        0
#> 47    RCCH           0    C          9        0
#> 48     RCN           0    H         12        0
#> 49    RNH2           0    C         10        0
#> 50    R2NH           0    H         16        0
#> 51     R3N           0    C         10        0
#> 52   ROPO3           0    H         16        0
#> 53     ROH           0    C         10        0
#> 54    RCHO           0    H         16        0
#> 55    RCOR           0    C         10        0
#> 56   RCOOH           0    H         16        0
#> 57   RCOOR           0    C         10        0
#> 58     ROR           0    H         16        0
#> 59    RCCH           0    C         10        0
#> 60     RCN           0    H         16        0
#> 61    RNH2           0    C         10        0
#> 62    R2NH           0    H         14        0
#> 63     R3N           0    O          1        0
#> 64   ROPO3           0    C         10        0
#> 65     ROH           1    H         14        0
#> 66    RCHO           0    O          1        0
#> 67    RCOR           0    C         10        0
#> 68   RCOOH           0    H         14        0
#> 69   RCOOR           0    O          1        0
#> 70     ROR           0    C         10        0
#> 71    RCCH           0    H         14        0
#> 72     RCN           0    O          1        0
#> 73    RNH2           0    C         10        0
#> 74    R2NH           0    H         14        0
#> 75     R3N           0    C         10        0
#> 76   ROPO3           0    H         14        0
#> 77     ROH           0    C         10        0
#> 78    RCHO           0    H         14        0
#> 79    RCOR           0    C         10        0
#> 80   RCOOH           0    H         14        0
#> 81   RCOOR           0    C         10        0
#> 82     ROR           0    H         14        0
#> 83    RCCH           0    C         10        0
#> 84     RCN           0    H         14        0
#> 85    RNH2           0    C         10        0
#> 86    R2NH           0    H         14        0
#> 87     R3N           0    C         10        0
#> 88   ROPO3           0    H         14        0
#> 89     ROH           0    C         10        0
#> 90    RCHO           0    H         14        0
#> 91    RCOR           0    C         10        0
#> 92   RCOOH           0    H         14        0
#> 93   RCOOR           0    C         10        0
#> 94     ROR           0    H         14        0
#> 95    RCCH           0    C         10        0
#> 96     RCN           0    H         14        0
#> 97    RNH2           0    C         10        0
#> 98    R2NH           0    H         14        0
#> 99     R3N           0    C         10        0
#> 100  ROPO3           0    H         14        0
#> 101    ROH           0    C         10        0
#> 102   RCHO           0    H         14        0
#> 103   RCOR           0    C         10        0
#> 104  RCOOH           0    H         14        0
#> 105  RCOOR           0    C         10        0
#> 106    ROR           0    H         14        0
#> 107   RCCH           0    C         10        0
#> 108    RCN           0    H         14        0
#> 109   RNH2           0    C         10        0
#> 110   R2NH           0    H         12        0
#> 111    R3N           0    C         10        0
#> 112  ROPO3           0    H         12        0
#> 113    ROH           0    C         10        0
#> 114   RCHO           0    H         12        0
#> 115   RCOR           0    C         10        0
#> 116  RCOOH           0    H         12        0
#> 117  RCOOR           0    C         10        0
#> 118    ROR           0    H         12        0
#> 119   RCCH           0    C         10        0
#> 120    RCN           0    H         12        0
#> 121   RNH2           0    C         10        0
#> 122   R2NH           0    H         12        0
#> 123    R3N           0    C         10        0
#> 124  ROPO3           0    H         12        0
#> 125    ROH           0    C         10        0
#> 126   RCHO           0    H         12        0
#> 127   RCOR           0    C         10        0
#> 128  RCOOH           0    H         12        0
#> 129  RCOOR           0    C         10        0
#> 130    ROR           0    H         12        0
#> 131   RCCH           0    C         10        0
#> 132    RCN           0    H         12        0
#> 133   RNH2           0    C         10        0
#> 134   R2NH           0    H         18        0
#> 135    R3N           0    O          1        0
#> 136  ROPO3           0    C         10        0
#> 137    ROH           1    H         18        0
#> 138   RCHO           0    O          1        0
#> 139   RCOR           0    C         10        0
#> 140  RCOOH           0    H         18        0
#> 141  RCOOR           0    O          1        0
#> 142    ROR           0    C         10        0
#> 143   RCCH           0    H         18        0
#> 144    RCN           0    O          1        0
#> 145   RNH2           0    C          9        0
#> 146   R2NH           0    H         18        0
#> 147    R3N           0    O          1        0
#> 148  ROPO3           0    C          9        0
#> 149    ROH           0    H         18        0
#> 150   RCHO           1    O          1        0
#> 151   RCOR           0    C          9        0
#> 152  RCOOH           0    H         18        0
#> 153  RCOOR           0    O          1        0
#> 154    ROR           0    C          9        0
#> 155   RCCH           0    H         18        0
#> 156    RCN           0    O          1        0
#> 157   RNH2           0    C         14        0
#> 158   R2NH           0    H         22        0
#> 159    R3N           0    O          2        0
#> 160  ROPO3           0    S          1        0
#> 161    ROH           0    C         14        0
#> 162   RCHO           0    H         22        0
#> 163   RCOR           0    O          2        0
#> 164  RCOOH           1    S          1        0
#> 165  RCOOR           0    C         14        0
#> 166    ROR           0    H         22        0
#> 167   RCCH           0    O          2        0
#> 168    RCN           0    S          1        0
#> 169   RNH2           0    C          2        0
#> 170   R2NH           0   Cl          2        0
#> 171    R3N           0    H          2        0
#> 172  ROPO3           0    O          1        0
#> 173    ROH           0    C          2        0
#> 174   RCHO           1   Cl          2        0
#> 175   RCOR           0    H          2        0
#> 176  RCOOH           0    O          1        0
#> 177  RCOOR           0    C          2        0
#> 178    ROR           0   Cl          2        0
#> 179   RCCH           0    H          2        0
#> 180    RCN           0    O          1        0
#> 181   RNH2           0    C         12        0
#> 182   R2NH           0    H         20        0
#> 183    R3N           0    O          2        0
#> 184  ROPO3           0    C         12        0
#> 185    ROH           0    H         20        0
#> 186   RCHO           0    O          2        0
#> 187   RCOR           0    C         12        0
#> 188  RCOOH           0    H         20        0
#> 189  RCOOR           1    O          2        0
#> 190    ROR           0    C         12        0
#> 191   RCCH           0    H         20        0
#> 192    RCN           0    O          2        0
#> 193   RNH2           0    C         10        0
#> 194   R2NH           0    H         16        0
#> 195    R3N           0    C         10        0
#> 196  ROPO3           0    H         16        0
#> 197    ROH           0    C         10        0
#> 198   RCHO           0    H         16        0
#> 199   RCOR           0    C         10        0
#> 200  RCOOH           0    H         16        0
#> 201  RCOOR           0    C         10        0
#> 202    ROR           0    H         16        0
#> 203   RCCH           0    C         10        0
#> 204    RCN           0    H         16        0
#> 
#> $FunctionalGroups
#>    Type.A Type.B Type.C Type.D Type.E
#> 1      No     No      ~    Yes    Yes
#> 2      No     No    Yes      ~    Yes
#> 3      No      ~     No     No     No
#> 4      No      ~     No     No     No
#> 5      No     No     No      ~      ~
#> 6      No      ~     No     No     No
#> 7      No      ~     No     No     No
#> 8      No      ~     No     No     No
#> 9      No     No     No      ~      ~
#> 10     No      ~     No     No     No
#> 11     No      ~     No     No     No
#> 12     No     No     No     No     No
#> 13     No     No    Yes      ~    Yes
#> 14     No     No    Yes      ~    Yes
#> 15     No     No     No     No     No
#> 16     No     No     No     No     No
#> 17     No     No     No     No     No
#>                                            Chemical
#> 1                                        3-octanone
#> 2                                            decane
#> 3                            1,3,5-trimethylbenzene
#> 4                            1,2,4-trimethylbenzene
#> 5      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene
#> 6                                2-phenyl-1-butanol
#> 7                                1,4-diethylbenzene
#> 8                                1,2-diethylbenzene
#> 9   1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene
#> 10                       2-methylprop-1-enylbenzene
#> 11                          [(E)-but-1-enyl]benzene
#> 12                   3,7-dimethyl-3-octa-1,6-dienol
#> 13                                          nonanal
#> 14               5-nonyl-2-thiophenecarboxylic acid
#> 15                         2,2-dichloroacetaldehyde
#> 16 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester
#> 17                    3,7-dimethylocta-1,3,6-triene
#> 
#> $BestChemMatch
#>    Type.A Type.B Type.C Type.D Type.E
#> 1      No     No     No   CMP7   CMP1
#> 2      No     No   CMP1     No   CMP1
#> 3      No     No     No     No     No
#> 4      No     No     No     No     No
#> 5      No     No     No     No     No
#> 6      No     No     No     No     No
#> 7      No     No     No     No     No
#> 8      No     No     No     No     No
#> 9      No     No     No     No     No
#> 10     No     No     No     No     No
#> 11     No     No     No     No     No
#> 12     No     No     No     No     No
#> 13     No     No   CMP1     No   CMP1
#> 14     No     No   CMP1     No   CMP1
#> 15     No     No     No     No     No
#> 16     No     No     No     No     No
#> 17     No     No     No     No     No
#>                                            Chemical
#> 1                                        3-octanone
#> 2                                            decane
#> 3                            1,3,5-trimethylbenzene
#> 4                            1,2,4-trimethylbenzene
#> 5      (4R)-1-methyl-4-(1-methylethenyl)cyclohexene
#> 6                                2-phenyl-1-butanol
#> 7                                1,4-diethylbenzene
#> 8                                1,2-diethylbenzene
#> 9   1-methyl-4-(1-methylethenyl)cyclohexa-1,3-diene
#> 10                       2-methylprop-1-enylbenzene
#> 11                          [(E)-but-1-enyl]benzene
#> 12                   3,7-dimethyl-3-octa-1,6-dienol
#> 13                                          nonanal
#> 14               5-nonyl-2-thiophenecarboxylic acid
#> 15                         2,2-dichloroacetaldehyde
#> 16 acetic acid 3,7-dimethylocta-1,6-dien-3-yl ester
#> 17                    3,7-dimethylocta-1,3,6-triene
#>