mzExacto

Uses the output from `spreadOut()` and a list of chemicals to extract data for. Samples that contain a chemical have all identified area(s) aggregated. The most likely identifications are determined by prioritizing matches of exact chemical names, followed by m/z overlaps within precise retention time windows that are determined by molecular masses published on PubChem.

mzExacto(data_in, chemicals, decontaminate = T)

Arguments

data_in

the list output from `spreadOut()`

chemicals

A vector containing chemical names in IUPAC notation

decontaminate

A logical set to remove uncertain chemicals or junk from output

Value

A data frame containing chemical names(`chemicals`), their optimal retention time, exact mass, best identified match factor, and aggregated component area across every sample it was identified in.

Details

Communicates with PubChem to collect information on every search chemical. Uses this information to search the raw `spreadOut()` data for samples where a chemical exists.

Examples

query_chemicals = c("Ethyl hexanoate","Methyl salicylate","Octanal","Undecane")
mzExacto(standard_spread, query_chemicals)
#> Preparing the data. Stay tuned. 
#> Acquiring relevant info from Pubchem [pubchem.ncbi.nlm.nih.gov]. Please be patient!!
#> [Current/Total] |--Exact Mass--|--CMP Name 1--|--CMP Name 2--|--Top MZ Peaks--|
#> [1 / 4]: 144.115029749 | hexanoic acid ethyl ester | Ethyl hexanoate | 88 | 79.07 | 43 | 29 | 99 | 60 | 53.52
#> [2 / 4]: 152.047344113 | 2-hydroxybenzoic acid methyl ester | Methyl salicylate | 120 | 92 | 64.72 | 152 | 66.62 | 121 | 65 | 30.11 | 30.88
#> [3 / 4]: 128.120115130 | octanal | Octanal | 43 | 41 | 10.41 | 44 | 98.48 | 66.79 | 79.38 | 56 | 114 | 29 | 57 | 52.64 | 17.62 | 96 | 87 | 100
#> [4 / 4]: 156.187800766 | undecane | Undecane | 43 | 11 | 57 | 15 | 71 | 155 | 51 | 45.96 | 94.31 | 29 | 48 | 32.03 | 85 | 99 | 55
#> Working on Octanal
#> Working on Ethyl hexanoate
#> Working on Methyl salicylate
#> Working on Undecane
#>            Compound          Mass               RT       Best Match
#> 1           Octanal 128.120115130  5.4620897525496 99.3245676204613
#> 2   Ethyl hexanoate 144.115029749 5.37971887422041  99.350118110327
#> 3 Methyl salicylate 152.047344113 8.29568988791984  98.161520884759
#> 4          Undecane 156.187800766 6.12919146738922 98.6771852019613
#>      Std_soln_00.D    Std_soln_07.D   Std_soln_00a.D
#> 1 379178.886500925 30943.1138493066 125725.898168008
#> 2 263866.042668436 9896.48814871315 294869.135678944
#> 3 30332.9022085754 6476.47845067295 209705.187766455
#> 4 86270.0501904642 243.912373077048 238776.228691087