standardifyIt

Post-merge function to quantify compound emission rates relative to internal or external standard(s). If using internal standard (IS), will use the contained `standardify()` function with user-specified inputs. If using external standard (ES), an input "matrix" from which standard curves can be derived.

standardifyIt(
  data_in,
  standard_type = "Internal",
  standard_used = "Tetradecane",
  IS_ng = 190.5,
  IS_uL = 1,
  collect_time = 1,
  sample_amt = 1,
  ES_calibration = NA
)

Arguments

data_in

mzExacto output

standard_type

specifies type of standardization to perform ("Internal" or "External")

standard_used

specifies the standard used as an internal standard (name your chemical)

IS_ng

specifies the quantified number of molecules in the standard

IS_uL

specifies the amount of standard added to the samples

collect_time

specifies how long samples were collected for (e.g. hour(s), day(s), year(s))

sample_amt

specifies how many individuals/items the samples were collected from

ES_calibration

input matrix from which calibrations will be fit for external standardization

Value

Returns the original data standardized relative to the user-specified internal or external inputs

Examples

standardifyIt(standard_exacto, standard_type = "Internal", standard_used = "Octanal",
IS_ng = 1, IS_uL = 1, collect_time = 1)
#>                     Compound          Mass               RT       Best Match
#> 2 Hexanoic acid, ethyl ester 144.115029749 5.37971887422041  99.350118110327
#> 3          Methyl salicylate 152.047344113 8.29568988791984  98.161520884759
#> 4                   Undecane 156.187800766 6.79854308486045 98.6771852019613
#>               Std_soln_00.D            Std_soln_07.D          Std_soln_00a.D
#> 2 0.69588801503005714987893 0.3198284502622840697050 2.345333300263317433121
#> 3 0.07999628483665954803428 0.2093027379924816688828 1.667955376116901033967
#> 4 0.22751807461266371213071 0.0000000000000000000000 1.899180933843951368800
standardifyIt(standard_exacto, standard_type = "External", ES_calibration = ExternalStandard_data)
#>                     Compound          Mass               RT       Best Match
#> 1                    Octanal 128.120115130  5.4620897525496 99.3245676204613
#> 2 Hexanoic acid, ethyl ester 144.115029749 5.37971887422041  99.350118110327
#> 3          Methyl salicylate 152.047344113 8.29568988791984  98.161520884759
#> 4                   Undecane 156.187800766 6.79854308486045 98.6771852019613
#>             Std_soln_00.D           Std_soln_07.D          Std_soln_00a.D
#> 1 1.619650165564917809746 1.620477842952729563564 1.620252523903682240203
#> 2 1.619924191194201057442 1.620527879556545647333 1.619850511912623414190
#> 3 1.620479293658454666982 1.620536010491357314933 1.620052913260099147763
#> 4 1.620346315127551228841                      NA 1.619983819939409208644